All Files for the game
These are all the files and code associated with our safari game. I have also included a release/executable file of the game which can be played in any linux system. This is a beginner project that my group mates and I created while learning C++ and GUI design and i'm glad to say that we definitely learned the power of C++ along the way. Any feedback would be great!!